requirements.txt
#pypi.org
django==5.0.7
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
django-storages==1.14.3
psycopg==3.1.19

mysqlclient==2.2.4



#pip install -r .\requirements.txt